fix: replace InMemorySessionStorage with SessionStorage and InMemoryStorage
