+++ packages/core/examples/metrics-example.ts
@@ -3,11 +3,13 @@ import {
   AIPexAgent,
   aisdk,
   ConversationManager,
-  InMemorySessionStorage,
+  InMemoryStorage,
+  type SerializedSession,
+  SessionStorage,
 } from "../src/index.js";
 
 async function demonstrateMetrics() {
-  const storage = new InMemorySessionStorage();
+  const storage = new SessionStorage(new InMemoryStorage<SerializedSession>());
   const manager = new ConversationManager(storage);
 
   const agent = AIPexAgent.create({
